refactor: remove lib/ directory from intelligence domain, follow codebase conventions
Move skills/, memories/, memory-v2/, and client.ts from
domains/intelligence/lib/ to the domain root. Update all imports.
Remove empty hooks/ directory.
