+++ apps/web/src/domains/intelligence/client.ts
