+++ apps/web/src/domains/intelligence/components/apps/app-viewer-container.tsx
@@ -2,7 +2,7 @@ import { useEffect, useMemo, useRef } from "react";
 
 import { AppNavBar } from "@/components/app-nav-bar.js";
 import { FETCH_PROXY_ALLOWED_PATH_RE, injectBridge } from "@/domains/chat/lib/app-bridge.js";
-import { client } from "@/domains/intelligence/lib/client.js";
+import { client } from "@/domains/intelligence/client.js";
 
 export interface AppViewerContainerProps {
   appId: string;
