+++ apps/web/src/domains/intelligence/components/identity-tab.tsx
@@ -11,8 +11,8 @@ import { useAssistantAvatar } from "@/domains/avatar/use-assistant-avatar.js";
 import type { AssistantIdentity } from "@/domains/chat/lib/api.js";
 import { fetchAssistantIdentity } from "@/domains/chat/lib/api.js";
 import type { CharacterComponents, CharacterTraits } from "@/domains/avatar/types.js";
-import { fetchSkills, installSkill, uninstallSkill } from "@/domains/intelligence/lib/skills/api.js";
-import type { SkillInfo } from "@/domains/intelligence/lib/skills/types.js";
+import { fetchSkills, installSkill, uninstallSkill } from "@/domains/intelligence/skills/api.js";
+import type { SkillInfo } from "@/domains/intelligence/skills/types.js";
 import { getAssistant } from "@/domains/assistant/api.js";
 
 export interface IdentityCardProps {
