+++ apps/web/src/domains/intelligence/components/skills/skill-origin-badge.tsx
@@ -2,7 +2,7 @@ import { Globe, Package, Puzzle, Terminal, User } from "lucide-react";
 import { createElement } from "react";
 
 import { Tag } from "@vellum/design-library";
-import type { SkillOrigin } from "@/domains/intelligence/lib/skills/types.js";
+import type { SkillOrigin } from "@/domains/intelligence/skills/types.js";
 
 const ORIGIN_META: Record<SkillOrigin, { label: string; icon: typeof Globe }> = {
   vellum: { label: "Vellum", icon: Package },
