+++ apps/web/src/domains/intelligence/memories/api.ts
@@ -12,7 +12,7 @@ import {
   client,
   extractErrorMessage,
   SDK_BASE_OPTIONS,
-} from "@/domains/intelligence/lib/client.js";
+} from "@/domains/intelligence/client.js";
 
 import type { MemoryItem, MemoryItemsListResponse } from "./types.js";
