+++ apps/web/src/domains/intelligence/memory-v2/api.ts
@@ -11,7 +11,7 @@ import {
   client,
   extractErrorMessage,
   SDK_BASE_OPTIONS,
-} from "@/domains/intelligence/lib/client.js";
+} from "@/domains/intelligence/client.js";
 
 import type { ConceptPageSummary, ListConceptPagesResult } from "./types.js";
