+++ apps/web/src/domains/intelligence/skills/category.ts
