+++ apps/web/src/domains/intelligence/skills/types.ts
